+++ src/variants/packed_color_array_yaml.cpp
@@ -81,9 +81,11 @@ Variant PackedColorArrayVariantConverter::decode(const ryml::ConstNodeRef &node,
 				context->pop_style();
 			}
 		} catch (const YAMLException &e) {
-			throw YAMLException(vformat("Failed to decode PackedColorArray value at index %d: %s", i, e.what()), e.get_location());
+			// NOTE: Cast size_t to int64_t for ARM64 macOS Variant compatibility
+			throw YAMLException(vformat("Failed to decode PackedColorArray value at index %d: %s", static_cast<int64_t>(i), e.what()), e.get_location());
 		} catch (const std::exception &e) {
-			throw YAMLException(vformat("Failed to decode PackedColorArray value at index %d: %s", i, e.what()), context->get_ryml_parser()->location(node[i]));
+			// NOTE: Cast size_t to int64_t for ARM64 macOS Variant compatibility
+			throw YAMLException(vformat("Failed to decode PackedColorArray value at index %d: %s", static_cast<int64_t>(i), e.what()), context->get_ryml_parser()->location(node[i]));
 		}
 	}
