+++ src/variants/packed_string_array_yaml.cpp
@@ -90,9 +90,11 @@ Variant PackedStringArrayVariantConverter::decode(const ryml::ConstNodeRef &node
 				}
 			}
 		} catch (const YAMLException &e) {
-			throw YAMLException(vformat("Failed to decode PackedStringArray value at index %d: %s", i, e.what()), e.get_location());
+			// NOTE: Cast size_t to int64_t for ARM64 macOS Variant compatibility
+			throw YAMLException(vformat("Failed to decode PackedStringArray value at index %d: %s", static_cast<int64_t>(i), e.what()), e.get_location());
 		} catch (const std::exception &e) {
-			throw YAMLException(vformat("Failed to decode PackedStringaArray value at index %d: %s", i, e.what()), context->get_ryml_parser()->location(node[i]));
+			// NOTE: Cast size_t to int64_t for ARM64 macOS Variant compatibility
+			throw YAMLException(vformat("Failed to decode PackedStringaArray value at index %d: %s", static_cast<int64_t>(i), e.what()), context->get_ryml_parser()->location(node[i]));
 		}
 	}
